moved selectors from place-screen in to child components
