+++ project/src/components/auth-form/auth-form.tsx
@@ -5,12 +5,11 @@ import { selectCity } from '../../store/places-process/places-process';
 
 import { CITIES } from '../../consts';
 
-
 function AuthForm(): JSX.Element {
   const dispatch = useAppDispatch();
   const [userData, setUserData] = useState({ email: '', password: '' });
 
-  function handleOnChange(e: React.ChangeEvent<HTMLInputElement>) {
+  function handleChangeInput(e: React.ChangeEvent<HTMLInputElement>) {
     const { value, name } = e.target;
 
     setUserData({
@@ -19,7 +18,7 @@ function AuthForm(): JSX.Element {
     });
   }
 
-  function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
+  function handleSubmitForm(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
 
     dispatch(loginAction(userData));
@@ -27,14 +26,14 @@ function AuthForm(): JSX.Element {
   }
 
   return (
-    <form onSubmit={handleSubmit} className="login__form form" action="#" method="post">
+    <form onSubmit={handleSubmitForm} className="login__form form" action="#" method="post">
       <div className="login__input-wrapper form__input-wrapper">
         <label className="visually-hidden">E-mail</label>
-        <input onChange={handleOnChange} value={userData.email} className="login__input form__input" type="email" name="email" placeholder="Email" required />
+        <input onChange={handleChangeInput} value={userData.email} className="login__input form__input" type="email" name="email" placeholder="Email" required />
       </div>
       <div className="login__input-wrapper form__input-wrapper">
         <label className="visually-hidden">Password</label>
-        <input onChange={handleOnChange} value={userData.password} className="login__input form__input" type="password" name="password" placeholder="Password" required />
+        <input onChange={handleChangeInput} value={userData.password} className="login__input form__input" type="password" name="password" placeholder="Password" required />
       </div>
       <button className="login__submit form__submit button" type="submit">Sign in</button>
     </form>
