+++ project/src/components/places-content/places-content.tsx
@@ -10,7 +10,6 @@ type PlacesContentProps = {
 };
 
 function PlacesContent({ places, currentCity }: PlacesContentProps): JSX.Element {
-
   return (
     <section className="cities__places places">
       <h2 className="visually-hidden">Places</h2>
